Saving progress on programming hobby
Not sure how to access hobby info or if there even is hobby info that exists so I started programming some if statementss to take care of hobby name and hobby icon. Had to rearrange some of the html elements to add div's to contain the icons within the hobby li's second span.

Need to finish hobby name programming and add in icons too.

Also added some styling to improve comparison to mockup